requirements.txt
# Requirements for Vehicle Plate Detection
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
eventlet==0.35.1

psutil==5.9.6
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.3

Pillow==10.1.0
opencv-python-headless==4.8.1.78

easyocr==1.7.1
pytesseract==0.3.10

transformers==4.36.0
datasets==2.15.0
accelerate==0.25.0
safetensors==0.4.1

huggingface-hub==0.20.2
hf-transfer==0.1.4

scikit-image==0.22.0
scipy==1.11.4
tqdm==4.66.1

ultralytics==8.0.196
# Core Flask & Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
blinker==1.9.0
eventlet==0.35.1
greenlet==3.3.0
python-socketio==5.10.0
python-engineio==4.13.0

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # For PostgreSQL (optional)
redis==5.0.1  # For Redis caching (optional)
celery==5.3.4  # For background tasks (optional)

# Computer Vision & Image Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
imageio==2.37.2

# OCR & Text Recognition
easyocr==1.7.1
pytesseract==0.3.10
pyclipper==1.4.0
shapely==2.0.2
python-bidi==0.6.7
reportlab==4.0.7  # For PDF support

# Machine Learning & Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
ultralytics==8.0.196
transformers==4.36.0
datasets==2.15.0
accelerate==0.25.0
safetensors==0.4.1
timm==0.9.7
onnx==1.20.1
onnxruntime==1.23.2

# Google Cloud Vision
google-cloud-vision==3.4.5
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.3.0
google-api-core==2.29.0
grpcio==1.76.0
grpcio-status==1.62.3
proto-plus==1.27.0

# Data Processing & Scientific Computing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
sympy==1.14.0

# Hugging Face
huggingface-hub==0.20.2
tokenizers==0.15.2
protobuf==4.25.8

# Utilities
psutil==5.9.6
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0
urllib3==2.6.3
certifi==2026.1.4
pyyaml==6.0.3
pyjwt==2.8.0
loguru==0.7.2
watchdog==3.0.0
humanfriendly==10.0
coloredlogs==15.0.1
wget==3.2

# File Handling & Formats
openpyxl==3.1.2
pandas==2.1.3
python-multipart==0.0.6
aiofiles==23.2.1
python-magic==0.4.27

# System & Monitoring
prometheus-client==0.19.0
prometheus-flask-exporter==0.22.4

# Testing & Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.2.0
pytest-asyncio==0.21.1
coverage==7.13.1
black==23.11.0
flake8==6.1.0
mypy-extensions==1.1.0

# Others
typing-extensions==4.15.0
pydantic==2.5.0
pydantic-settings==2.1.0
colorama==0.4.6
click==8.3.1
joblib==1.3.2
networkx==3.6.1
tifffile==2026.1.14
flatbuffers==25.12.19
ml-dtypes==0.5.4